+++ backend/Muhasebe/Common/Constants/MuhasebeFisDurumlari.cs
@@ -1,0 +1,17 @@
+namespace STYS.Muhasebe.Common.Constants;
+
+public static class MuhasebeFisDurumlari
+{
+    public const string Taslak = "Taslak";
+    public const string Onayli = "Onayli";
+    public const string Iptal = "Iptal";
+    public const string TersKayit = "TersKayit";
+
+    public static readonly string[] Hepsi =
+    [
+        Taslak,
+        Onayli,
+        Iptal,
+        TersKayit
+    ];
+}